requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
numpy
Pillow
tensorflow
opencv-python
torchvision
torch
pandas
tqdm
facenet_pytorch
scikit-image
scikit-learn
matplotlib
seaborn
timm